Renamed JAPI class to FrontController
