+++ tests/integration/FrontControllerTest.php
@@ -18,7 +18,7 @@
 
 declare(strict_types=1);
 
-namespace gordonmcvey\WarpCore\test\unit;
+namespace gordonmcvey\WarpCore\test\integration;
 
 use gordonmcvey\httpsupport\enum\statuscodes\ClientErrorCodes;
 use gordonmcvey\httpsupport\request\RequestInterface;
@@ -27,18 +27,19 @@ use gordonmcvey\httpsupport\response\sender\ResponseSenderInterface;
 use gordonmcvey\WarpCore\Exceptions\AccessDenied;
 use gordonmcvey\WarpCore\Exceptions\Auth;
 use gordonmcvey\WarpCore\Exceptions\Routing;
+use gordonmcvey\WarpCore\FrontController;
 use gordonmcvey\WarpCore\interface\controller\RequestHandlerInterface;
 use gordonmcvey\WarpCore\interface\error\ErrorHandlerInterface;
-use gordonmcvey\WarpCore\JAPI;
-use gordonmcvey\WarpCore\middleware\CallStack;
+use gordonmcvey\WarpCore\interface\middleware\MiddlewareInterface;
+use gordonmcvey\WarpCore\interface\middleware\MiddlewareProviderInterface;
 use gordonmcvey\WarpCore\middleware\CallStackFactory;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\MockObject\Exception;
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use Psr\Log\LoggerInterface;
 use RuntimeException;
 
-class JAPITest extends TestCase
+class FrontControllerTest extends TestCase
 {
     /**
      * @throws Exception
@@ -46,32 +47,17 @@ class JAPITest extends TestCase
     #[Test]
     public function itHandlesATypicalDispatchCycle(): void
     {
-        $mockCallStackFactory = $this->createMock(CallStackFactory::class);
-        $mockCallStack = $this->createMock(CallStack::class);
         $mockController = $this->createMock(RequestHandlerInterface::class);
         $mockRequest = $this->createMock(RequestInterface::class);
         $mockResponse = $this->createMock(ResponseInterface::class);
         $mockErrorHandler = $this->createMock(ErrorHandlerInterface::class);
         $mockSender = $this->createMock(ResponseSenderInterface::class);
 
-        $japi = new JAPI($mockCallStackFactory, $mockErrorHandler, $mockSender);
-
-        $mockCallStackFactory->expects($this->once())
-            ->method("make")
-            ->with($mockController, $japi)
-            ->willReturn($mockCallStack)
-        ;
-
         $mockController->expects($this->once())
             ->method("dispatch")
             ->with($mockRequest)
             ->willReturn($mockResponse)
         ;
-
-        $mockCallStack->expects($this->once())
-            ->method("dispatch")
-            ->with($mockRequest)
-            ->willReturn($mockController->dispatch($mockRequest));
 
         $mockErrorHandler->expects($this->never())
             ->method("handle")
@@ -83,6 +69,7 @@ class JAPITest extends TestCase
             ->willReturnSelf()
         ;
 
+        $japi = new FrontController(new CallStackFactory(), $mockErrorHandler, $mockSender);
         $japi->bootstrap($mockController, $mockRequest);
     }
 
@@ -90,34 +77,19 @@ class JAPITest extends TestCase
      * @throws Exception
      */
     #[Test]
-    public function itHandlesATypicalDispatchCycleWithControllerFactoryFunction(): void
+    public function itHandlesATypicalDispatchCycleWithFactoryFunction(): void
     {
-        $mockCallStackFactory = $this->createMock(CallStackFactory::class);
-        $mockCallStack = $this->createMock(CallStack::class);
         $mockController = $this->createMock(RequestHandlerInterface::class);
         $mockRequest = $this->createMock(RequestInterface::class);
         $mockResponse = $this->createMock(ResponseInterface::class);
         $mockErrorHandler = $this->createMock(ErrorHandlerInterface::class);
         $mockSender = $this->createMock(ResponseSenderInterface::class);
 
-        $japi = new JAPI($mockCallStackFactory, $mockErrorHandler, $mockSender);
-
-        $mockCallStackFactory->expects($this->once())
-            ->method("make")
-            ->with($mockController, $japi)
-            ->willReturn($mockCallStack)
-        ;
-
         $mockController->expects($this->once())
             ->method("dispatch")
             ->with($mockRequest)
             ->willReturn($mockResponse)
         ;
-
-        $mockCallStack->expects($this->once())
-            ->method("dispatch")
-            ->with($mockRequest)
-            ->willReturn($mockController->dispatch($mockRequest));
 
         $mockErrorHandler->expects($this->never())
             ->method("handle")
@@ -129,6 +101,7 @@ class JAPITest extends TestCase
             ->willReturnSelf()
         ;
 
+        $japi = new FrontController(new CallStackFactory(), $mockErrorHandler, $mockSender);
         $japi->bootstrap(fn() => $mockController, $mockRequest);
     }
 
@@ -136,34 +109,19 @@ class JAPITest extends TestCase
      * @throws Exception
      */
     #[Test]
-    public function itHandlesATypicalDispatchCycleWithControllerFactoryObject(): void
+    public function itHandlesATypicalDispatchCycleWithFactoryObject(): void
     {
-        $mockCallStackFactory = $this->createMock(CallStackFactory::class);
-        $mockCallStack = $this->createMock(CallStack::class);
         $mockController = $this->createMock(RequestHandlerInterface::class);
         $mockRequest = $this->createMock(RequestInterface::class);
         $mockResponse = $this->createMock(ResponseInterface::class);
         $mockErrorHandler = $this->createMock(ErrorHandlerInterface::class);
         $mockSender = $this->createMock(ResponseSenderInterface::class);
 
-        $japi = new JAPI($mockCallStackFactory, $mockErrorHandler, $mockSender);
-
-        $mockCallStackFactory->expects($this->once())
-            ->method("make")
-            ->with($mockController, $japi)
-            ->willReturn($mockCallStack)
-        ;
-
         $mockController->expects($this->once())
             ->method("dispatch")
             ->with($mockRequest)
             ->willReturn($mockResponse)
         ;
-
-        $mockCallStack->expects($this->once())
-            ->method("dispatch")
-            ->with($mockRequest)
-            ->willReturn($mockController->dispatch($mockRequest));
 
         $mockErrorHandler->expects($this->never())
             ->method("handle")
@@ -176,6 +134,7 @@ class JAPITest extends TestCase
             ->willReturnSelf()
         ;
 
+        $japi = new FrontController(new CallStackFactory(), $mockErrorHandler, $mockSender);
         $japi->bootstrap(
             new readonly class ($mockController) {
                 public function __construct(private RequestHandlerInterface $controller)
@@ -186,7 +145,7 @@ class JAPITest extends TestCase
                     return $this->controller;
                 }
             },
-            $mockRequest,
+            $mockRequest
         );
     }
 
@@ -194,12 +153,139 @@ class JAPITest extends TestCase
      * @throws Exception
      */
     #[Test]
-    public function itHandlesABootstrappingError(): void
+    public function itHandlesATypicalDispatchCycleWithGlobalMiddleware(): void
     {
-        $mockCallStackFactory = $this->createMock(CallStackFactory::class);
+        $mockController = $this->createMock(RequestHandlerInterface::class);
         $mockRequest = $this->createMock(RequestInterface::class);
         $mockResponse = $this->createMock(ResponseInterface::class);
         $mockErrorHandler = $this->createMock(ErrorHandlerInterface::class);
+        $mockSender = $this->createMock(ResponseSenderInterface::class);
+
+        $mockController->expects($this->once())
+            ->method("dispatch")
+            ->with($mockRequest)
+            ->willReturn($mockResponse)
+        ;
+
+        $mockRequest->expects($this->once())
+            ->method("setHeader")
+            ->with("foo", "bar")
+            ->willReturnSelf()
+        ;
+
+        $mockResponse->expects($this->once())
+            ->method("setHeader")
+            ->with("baz", "quux")
+            ->willReturnSelf()
+        ;
+
+        $mockErrorHandler->expects($this->never())
+            ->method("handle")
+        ;
+
+        $middleware = new class implements MiddlewareInterface
+        {
+            public function handle(RequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
+            {
+                $request->setHeader("foo", "bar");
+                $response = $handler->dispatch($request);
+                $response->setHeader("baz", "quux");
+
+                return $response;
+            }
+        };
+
+        $mockSender->expects($this->once())
+            ->method("send")
+            ->with($mockResponse)
+            ->willReturnSelf()
+        ;
+
+        $japi = new FrontController(new CallStackFactory(), $mockErrorHandler, $mockSender);
+        $japi->addMiddleware($middleware)->bootstrap($mockController, $mockRequest);
+    }
+
+    /**
+     * @throws Exception
+     */
+    #[Test]
+    public function itHandlesATypicalDispatchCycleWithControllerMiddleware(): void
+    {
+        /** @var RequestHandlerInterface&MiddlewareProviderInterface&MockObject $mockController */
+        $mockController = $this->createMockForIntersectionOfInterfaces([
+            RequestHandlerInterface::class,
+            MiddlewareProviderInterface::class,
+        ]);
+        $mockRequest = $this->createMock(RequestInterface::class);
+        $mockResponse = $this->createMock(ResponseInterface::class);
+        $mockErrorHandler = $this->createMock(ErrorHandlerInterface::class);
+        $mockSender = $this->createMock(ResponseSenderInterface::class);
+
+        $mockRequest->expects($this->once())
+            ->method("setHeader")
+            ->with("foo", "bar")
+            ->willReturnSelf()
+        ;
+
+        $mockResponse->expects($this->once())
+            ->method("setHeader")
+            ->with("baz", "quux")
+            ->willReturnSelf()
+        ;
+
+        $middleware = new class implements MiddlewareInterface
+        {
+            public function handle(RequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
+            {
+                $request->setHeader("foo", "bar");
+                $response = $handler->dispatch($request);
+                $response->setHeader("baz", "quux");
+
+                return $response;
+            }
+        };
+
+        $mockController->expects($this->once())
+            ->method("getAllMiddleware")
+            ->willReturn([$middleware])
+        ;
+
+        $mockController->expects($this->once())
+            ->method("dispatch")
+            ->with($mockRequest)
+            ->willReturn($mockResponse)
+        ;
+
+        $mockErrorHandler->expects($this->never())
+            ->method("handle")
+        ;
+
+        $mockSender->expects($this->once())
+            ->method("send")
+            ->with($mockResponse)
+            ->willReturnSelf()
+        ;
+
+        $japi = new FrontController(new CallStackFactory(), $mockErrorHandler, $mockSender);
+        $japi->bootstrap($mockController, $mockRequest);
+    }
+
+    /**
+     * @throws Exception
+     */
+    #[Test]
+    public function itHandlesABootStrappingError(): void
+    {
+        $mockRequest = $this->createMock(RequestInterface::class);
+        $mockResponse = $this->createMock(ResponseInterface::class);
+        $mockErrorHandler = $this->createMock(ErrorHandlerInterface::class);
+        $mockSender = $this->createMock(ResponseSenderInterface::class);
+
+        $mockSender->expects($this->once())
+            ->method("send")
+            ->with($mockResponse)
+            ->willReturnSelf()
+        ;
 
         $mockErrorHandler->expects($this->once())
             ->method("handle")
@@ -207,17 +293,7 @@ class JAPITest extends TestCase
             ->willReturn($mockResponse)
         ;
 
-        $mockSender = $this->createMock(ResponseSenderInterface::class);
-
-        $japi = new JAPI($mockCallStackFactory, $mockErrorHandler, $mockSender);
-
-        // JAPI expectations
-        $mockSender->expects($this->once())
-            ->method("send")
-            ->with($mockResponse)
-            ->willReturnSelf()
-        ;
-
+        $japi = new FrontController(new CallStackFactory(), $mockErrorHandler, $mockSender);
         $japi->bootstrap(fn() => "Hello", $mockRequest);
     }
 
@@ -227,20 +303,10 @@ class JAPITest extends TestCase
     #[Test]
     public function itHandlesARoutingError(): void
     {
-        $mockCallStackFactory = $this->createMock(CallStackFactory::class);
         $mockRequest = $this->createMock(RequestInterface::class);
         $mockResponse = $this->createMock(ResponseInterface::class);
         $mockErrorHandler = $this->createMock(ErrorHandlerInterface::class);
-
-        $mockErrorHandler->expects($this->once())
-            ->method("handle")
-            ->with($this->isInstanceOf(Routing::class))
-            ->willReturn($mockResponse)
-        ;
-
         $mockSender = $this->createMock(ResponseSenderInterface::class);
-
-        $japi = new JAPI($mockCallStackFactory, $mockErrorHandler, $mockSender);
 
         $mockSender->expects($this->once())
             ->method("send")
@@ -248,6 +314,13 @@ class JAPITest extends TestCase
             ->willReturnSelf()
         ;
 
+        $mockErrorHandler->expects($this->once())
+            ->method("handle")
+            ->with($this->isInstanceOf(Routing::class))
+            ->willReturn($mockResponse)
+        ;
+
+        $japi = new FrontController(new CallStackFactory(), $mockErrorHandler, $mockSender);
         $japi->bootstrap(fn() => throw new Routing(), $mockRequest);
     }
 
@@ -257,32 +330,16 @@ class JAPITest extends TestCase
     #[Test]
     public function itHandlesAuthError(): void
     {
-        $mockCallStackFactory = $this->createMock(CallStackFactory::class);
-        $mockCallStack = $this->createMock(CallStack::class);
         $mockController = $this->createMock(RequestHandlerInterface::class);
         $mockRequest = $this->createMock(RequestInterface::class);
         $mockResponse = $this->createMock(ResponseInterface::class);
         $mockErrorHandler = $this->createMock(ErrorHandlerInterface::class);
         $mockSender = $this->createMock(ResponseSenderInterface::class);
 
-        $japi = new JAPI($mockCallStackFactory, $mockErrorHandler, $mockSender);
-
-        $mockCallStackFactory->expects($this->once())
-            ->method("make")
-            ->with($mockController, $japi)
-            ->willReturn($mockCallStack)
-        ;
-
-        $mockCallStack->expects($this->once())
+        $mockController->expects($this->once())
             ->method("dispatch")
             ->with($mockRequest)
             ->willThrowException(new Auth())
-        ;
-
-        $mockErrorHandler->expects($this->once())
-            ->method("handle")
-            ->with($this->isInstanceOf(Auth::class))
-            ->willReturn($mockResponse)
         ;
 
         $mockSender->expects($this->once())
@@ -291,6 +348,13 @@ class JAPITest extends TestCase
             ->willReturnSelf()
         ;
 
+        $mockErrorHandler->expects($this->once())
+            ->method("handle")
+            ->with($this->isInstanceOf(Auth::class))
+            ->willReturn($mockResponse)
+        ;
+
+        $japi = new FrontController(new CallStackFactory(), $mockErrorHandler, $mockSender);
         $japi->bootstrap($mockController, $mockRequest);
     }
 
@@ -300,32 +364,17 @@ class JAPITest extends TestCase
     #[Test]
     public function itHandlesAccessDeniedError(): void
     {
-        $mockCallStackFactory = $this->createMock(CallStackFactory::class);
-        $mockCallStack = $this->createMock(CallStack::class);
         $mockController = $this->createMock(RequestHandlerInterface::class);
         $mockRequest = $this->createMock(RequestInterface::class);
         $mockResponse = $this->createMock(ResponseInterface::class);
         $mockErrorHandler = $this->createMock(ErrorHandlerInterface::class);
         $mockSender = $this->createMock(ResponseSenderInterface::class);
 
-        $japi = new JAPI($mockCallStackFactory, $mockErrorHandler, $mockSender);
-
-        $mockCallStackFactory->expects($this->once())
-            ->method("make")
-            ->with($mockController, $japi)
-            ->willReturn($mockCallStack)
-        ;
-
-        $mockCallStack->expects($this->once())
+        $mockController->expects($this->once())
             ->method("dispatch")
             ->with($mockRequest)
             ->willThrowException(new AccessDenied())
         ;
-
-        $mockErrorHandler->expects($this->once())
-            ->method("handle")
-            ->with($this->isInstanceOf(AccessDenied::class))
-            ->willReturn($mockResponse);
 
         $mockSender->expects($this->once())
             ->method("send")
@@ -333,6 +382,13 @@ class JAPITest extends TestCase
             ->willReturnSelf()
         ;
 
+        $mockErrorHandler->expects($this->once())
+            ->method("handle")
+            ->with($this->isInstanceOf(AccessDenied::class))
+            ->willReturn($mockResponse)
+        ;
+
+        $japi = new FrontController(new CallStackFactory(), $mockErrorHandler, $mockSender);
         $japi->bootstrap($mockController, $mockRequest);
     }
 
@@ -342,32 +398,16 @@ class JAPITest extends TestCase
     #[Test]
     public function itHandlesGeneralError(): void
     {
-        $mockCallStackFactory = $this->createMock(CallStackFactory::class);
-        $mockCallStack = $this->createMock(CallStack::class);
         $mockController = $this->createMock(RequestHandlerInterface::class);
         $mockRequest = $this->createMock(RequestInterface::class);
         $mockResponse = $this->createMock(ResponseInterface::class);
         $mockErrorHandler = $this->createMock(ErrorHandlerInterface::class);
         $mockSender = $this->createMock(ResponseSenderInterface::class);
 
-        $japi = new JAPI($mockCallStackFactory, $mockErrorHandler, $mockSender);
-
-        $mockCallStackFactory->expects($this->once())
-            ->method("make")
-            ->with($mockController, $japi)
-            ->willReturn($mockCallStack)
-        ;
-
-        $mockCallStack->expects($this->once())
+        $mockController->expects($this->once())
             ->method("dispatch")
             ->with($mockRequest)
             ->willThrowException(new RuntimeException(code: 12345))
-        ;
-
-        $mockErrorHandler->expects($this->once())
-            ->method("handle")
-            ->with($this->isInstanceOf(RuntimeException::class))
-            ->willReturn($mockResponse)
         ;
 
         $mockSender->expects($this->once())
@@ -376,6 +416,13 @@ class JAPITest extends TestCase
             ->willReturnSelf()
         ;
 
+        $mockErrorHandler->expects($this->once())
+            ->method("handle")
+            ->with($this->isInstanceOf(RuntimeException::class))
+            ->willReturn($mockResponse)
+        ;
+
+        $japi = new FrontController(new CallStackFactory(), $mockErrorHandler, $mockSender);
         $japi->bootstrap($mockController, $mockRequest);
     }
 
@@ -385,32 +432,16 @@ class JAPITest extends TestCase
     #[Test]
     public function itHandlesGeneralErrorWithValidErrorCode(): void
     {
-        $mockCallStackFactory = $this->createMock(CallStackFactory::class);
-        $mockCallStack = $this->createMock(CallStack::class);
         $mockController = $this->createMock(RequestHandlerInterface::class);
         $mockRequest = $this->createMock(RequestInterface::class);
         $mockResponse = $this->createMock(ResponseInterface::class);
         $mockErrorHandler = $this->createMock(ErrorHandlerInterface::class);
         $mockSender = $this->createMock(ResponseSenderInterface::class);
 
-        $japi = new JAPI($mockCallStackFactory, $mockErrorHandler, $mockSender);
-
-        $mockCallStackFactory->expects($this->once())
-            ->method("make")
-            ->with($mockController, $japi)
-            ->willReturn($mockCallStack)
-        ;
-
-        $mockCallStack->expects($this->once())
+        $mockController->expects($this->once())
             ->method("dispatch")
             ->with($mockRequest)
             ->willThrowException(new RuntimeException(code: ClientErrorCodes::UNAVAILABLE_FOR_LEGAL_REASONS->value))
-        ;
-
-        $mockErrorHandler->expects($this->once())
-            ->method("handle")
-            ->with($this->isInstanceOf(RuntimeException::class))
-            ->willReturn($mockResponse)
         ;
 
         $mockSender->expects($this->once())
@@ -419,29 +450,13 @@ class JAPITest extends TestCase
             ->willReturnSelf()
         ;
 
-        $japi->bootstrap($mockController, $mockRequest);
-    }
-
-    /**
-     * @throws Exception
-     */
-    #[Test]
-    public function itLogsErrorsIfGivenALogger(): void
-    {
-        $mockCallStackFactory = $this->createMock(CallStackFactory::class);
-        $mockRequest = $this->createMock(RequestInterface::class);
-        $mockLogger = $this->createMock(LoggerInterface::class);
-        $mockErrorHandler = $this->createMock(ErrorHandlerInterface::class);
-        $mockSender = $this->createMock(ResponseSenderInterface::class);
-
-        $japi = new JAPI($mockCallStackFactory, $mockErrorHandler, $mockSender);
-
-        $mockLogger->expects($this->once())
-            ->method("error")
-            ->with("[JAPI] [500] Error: Unable to bootstrap")
+        $mockErrorHandler->expects($this->once())
+            ->method("handle")
+            ->with($this->isInstanceOf(RuntimeException::class))
+            ->willReturn($mockResponse)
         ;
 
-        $japi->setLogger($mockLogger);
-        $japi->bootstrap(fn() => "Hello", $mockRequest);
+        $japi = new FrontController(new CallStackFactory(), $mockErrorHandler, $mockSender);
+        $japi->bootstrap($mockController, $mockRequest);
     }
 }
